day picker

+++ core/src/main/java/com/aminovic/core/data/preferences/DefaultPreferences.kt
@@ -74,9 +74,9 @@ class DefaultPreferences(
         val genderString = sharedPref.getString(Preferences.KEY_GENDER, null)
         val activityLevelString = sharedPref.getString(Preferences.KEY_ACTIVITY_LEVEL, null)
         val goalTypeString = sharedPref.getString(Preferences.KEY_GOAL_TYPE, null)
-        val carbRatio = sharedPref.getFloat(Preferences.KEY_CARB_RATIO, -1f)
-        val proteinRatio = sharedPref.getFloat(Preferences.KEY_PROTEIN_RATIO, -1f)
-        val fatRatio = sharedPref.getFloat(Preferences.KEY_FAT_RATIO, -1f)
+        val carbRatio = sharedPref.getInt(Preferences.KEY_CARB_RATIO, -1)
+        val proteinRatio = sharedPref.getInt(Preferences.KEY_PROTEIN_RATIO, -1)
+        val fatRatio = sharedPref.getInt(Preferences.KEY_FAT_RATIO, -1)
 
         return UserInfo(
             gender = Gender.fromString(genderString ?: "male"),
