+++ core/src/main/java/com/aminovic/core/domain/model/UserInfo.kt
@@ -7,7 +7,7 @@ data class UserInfo(
     val height: Int,
     val activityLevel: ActivityLevel,
     val goalType: GoalType,
-    val carbRatio: Float,
-    val proteinRatio: Float,
-    val fatRatio: Float
+    val carbRatio: Int,
+    val proteinRatio: Int,
+    val fatRatio: Int
 )
